requirements.txt
python-telegram-bot==20.1
urllib3==1.26.15
